Workaround app load problem - Hardcode bootloader settings.
Ideally this setting in flash should be written by the bootloader
when a app is succesfully written in the flash. However it is
not happening now. So the bootloader stays in DFU mode all the
time.

For now hardcode bootloader settings to say there is a valid app so
that bootloader will load app.