Workaround APP->DFU transistion problem.
Currently when APP transfers control to DFU there is some
GATT error which causes the OTA to fail.

The only way to pass is the OTA has to be started when
the device is in DFU mode.  With this patch
when the OTA is second time tried it passes because
the device stays in DFU mode after first failure.