lyra: Create Button Service.
Lyra will provide a BLE button service(button pressed, held).

This is different from button config service. Button config service
provides a way to program what should happen when a button is
pressed, ie it is used configure what action should be taken by
lyra when a button is pressed.

This new button service just provides what is the current status of
the button. When written it will resets the button status. Also
it broadcasts button status as service data in advertisment packet.
Mobile app can use this to detect a button is pressed without
connecting to lyra. However it has to make a connection and write
to button service to reset the button service data.